refactor(home screen): improve layout and tap interactions for recommendations and features sections
