+++ lib/main.dart
@@ -30,7 +30,7 @@ import 'package:nutrimpasi/screens/food/food_list_screen.dart';
 import 'package:nutrimpasi/screens/forum/forum_screen.dart';
 import 'package:nutrimpasi/screens/home_screen.dart';
 import 'package:nutrimpasi/screens/setting/profile_screen.dart';
-import 'package:nutrimpasi/screens/schedule_screen.dart';
+import 'package:nutrimpasi/screens/features/schedule_screen.dart';
 import 'package:nutrimpasi/screens/splash_screen.dart';
 import 'screens/onboarding_screen.dart';
 import 'constants/colors.dart';
